Initial Commit
Currently Contains:
- Cheello's Voxel Doom v2.1 Vanilla Voxels
- Mickromash's HDest Voxels
- Hara's Peppergrinder Voxels

Cheello's Voxels are free to use with credit, whereas the ZScript that accompanies it is MIT.  Peppergrinder Voxels have been permission, and since HDest Voxels are mostly Cheello's anyways, this project will be licensed with an overall GPL v3 to cover the existing rights to their respective owners, but the original licenses where available will be provided as needed.

For now all original assets have been included, although those that are intended to be overridden for translation purposes (for instance Medikits & Stimpacks) will be renamed to prevent collisions.  Once they have been determined to not be needed they can be removed to prevent bloating the PK3 more than is necessary.